Add header and footer to needed pages, restructure project
