+++ src/Pages.js
@@ -1,10 +1,10 @@
 import LoginPage from "./pages/auth/LoginPage";
 import RegisterPage from "./pages/auth/RegisterPage";
 import ProfilePage from "./pages/user/ProfilePage";
-import ListingPage from "./pages/ListingPage";
+import ListingPage from "./pages/product/ListingPage";
 import HomePage from "./pages/HomePage";
-import FaqPage from "./pages/FaqPage";
-import AboutUsPage from "./pages/AboutUsPage";
+import FaqPage from "./pages/support/FaqPage";
+import AboutUsPage from "./pages/support/AboutUsPage";
 import ProductDetailPage from "./pages/product/ProductDetailPage";
 import CopyrightPage from "./pages/legal/CopyrightPage";
 import PrivacyPolicyPage from "./pages/legal/PrivacyPolicyPage";
