+++ src/pages/legal/CopyrightPage.js
@@ -3,7 +3,7 @@ import ToTopButton from "../../components/ToTopButton";
 
 const CopyrightPage = () => {
   return (
-    <Layout>
+    <Layout header footer>
       <ul>
         <li>
           <a>Home</a>
